+++ wai/ChangeLog.md
@@ -1,6 +1,6 @@
 ## 3.2.1.2
 
-* Remove dependency on blaze-builder
+* Remove dependency on blaze-builder [#683](https://github.com/yesodweb/wai/pull/683)
 
 ## 3.2.1.1