+++ wai/ChangeLog.md
@@ -1,3 +1,7 @@
+## 3.2.1.2
+
+* Remove dependency on blaze-builder [#683](https://github.com/yesodweb/wai/pull/683)
+
 ## 3.2.1.1
 
 * Relax upper bound on bytestring-builder